Fix some methods etc Rename 'input' module to 'io' Deprecating animator.py
